#144 Extract element provider into separate interface
